requirements.txt
asgiref==3.3.4
Django==3.2
django-countries==7.1
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.12.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Pillow==8.2.0
PyJWT==2.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
tqdm==4.60.0
utils==1.0.1
